refactor: move route data Types to their own module
