+++ neuron/src/lib/Neuron/Frontend/Route/Data.hs
@@ -9,29 +9,23 @@
 
 module Neuron.Frontend.Route.Data where
 
+import Data.Foldable (Foldable (maximum))
 import qualified Data.Set as Set
+import Data.TagTree
+import Data.Tree
 import Neuron.Cache.Type (NeuronCache (..))
 import qualified Neuron.Config.Type as Config
-import qualified Neuron.Frontend.Impulse as Impulse
 import Neuron.Frontend.Manifest (Manifest)
-import Neuron.Frontend.Route
-  ( HeadHtml,
-    Impulse,
-    SiteData (SiteData),
-    ZettelData (ZettelData),
-  )
+import Neuron.Frontend.Route.Data.Types
 import qualified Neuron.Frontend.Theme as Theme
 import Neuron.Zettelkasten.Connection (Connection (Folgezettel))
 import qualified Neuron.Zettelkasten.Graph as G
 import Neuron.Zettelkasten.ID (indexZid)
+import Neuron.Zettelkasten.Query
 import Neuron.Zettelkasten.Query.Eval
   ( buildQueryUrlCache,
   )
 import Neuron.Zettelkasten.Zettel
-  ( ZettelC,
-    ZettelT (..),
-    sansContent,
-  )
 import Relude hiding (traceShowId)
 import qualified Text.Pandoc.Util as P
 
@@ -50,9 +44,29 @@ mkZettelData NeuronCache {..} zC = do
       qurlcache = buildQueryUrlCache (G.getZettels _neuronCache_graph) allUrls
   ZettelData zC qurlcache upTree backlinks _neuronCache_graph
 
-mkImpulseData :: NeuronCache -> Impulse
+mkImpulseData :: NeuronCache -> ImpulseData
 mkImpulseData NeuronCache {..} =
-  Impulse.buildImpulse _neuronCache_graph _neuronCache_errors
+  buildImpulse _neuronCache_graph _neuronCache_errors
+  where
+    buildImpulse graph errors =
+      let (orphans, clusters) = partitionEithers $
+            flip fmap (G.categoryClusters graph) $ \case
+              [Node z []] -> Left z -- Orphans (cluster of exactly one)
+              x -> Right x
+          clustersWithUplinks :: [Forest (Zettel, [Zettel])] =
+            -- Compute backlinks for each node in the tree.
+            flip fmap clusters $ \(zs :: [Tree Zettel]) ->
+              G.backlinksMulti Folgezettel zs graph
+          stats = Stats (length $ G.getZettels graph) (G.connectionCount graph)
+          pinnedZettels = zettelsByTag (G.getZettels graph) $ mkDefaultTagQuery [mkTagPattern "pinned"]
+       in ImpulseData (fmap sortCluster clustersWithUplinks) orphans errors stats pinnedZettels
+    -- TODO: Either optimize or get rid of this (or normalize the sorting somehow)
+    sortCluster fs =
+      sortZettelForest $
+        flip fmap fs $ \Node {..} ->
+          Node rootLabel $ sortZettelForest subForest
+    -- Sort zettel trees so that trees containing the most recent zettel (by ID) come first.
+    sortZettelForest = sortOn (Down . maximum)
 
 mkSiteData :: NeuronCache -> HeadHtml -> Manifest -> SiteData
 mkSiteData NeuronCache {..} headHtml manifest =
